% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsem_summary.R
\name{path.networksem}
\alias{path.networksem}
\title{Calculate a mediation effect from a networksem model}
\usage{
path.networksem(res, predictor, mediator, outcome)
}
\arguments{
\item{res}{a networksem output file}

\item{predictor}{a character string of the predictor variable}

\item{mediator}{a character string of the mediator variable}

\item{outcome}{a character string of the outcome variable}
}
\value{
a target path
}
\description{
Calculate a mediation effect from a networksem model
}
