% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFun.R
\name{sem.text}
\alias{sem.text}
\title{TextSEM}
\usage{
sem.text(
  model,
  data,
  text_var,
  text_stats = c("OverallSenti"),
  polarity_dt = lexicon::hash_sentiment_jockers_rinker,
  valence_shifters_dt = lexicon::hash_valence_shifters,
  missing = "ML",
  fixed.x = FALSE,
  ...
)
}
\arguments{
\item{model}{The structural equation model.}

\item{data}{The input data frame.}

\item{text_var}{A vector of text variable(s).}

\item{text_stats}{A vector of text sentiment statistic(s). Currently only support one statistic - the overall sentiment.}
}
\value{
A list containing three items: the SEM with added sentiment variables, the input data frame with added text sentiment statistics, and the model estimates.
}
\description{
Get the sentiment of the text and use that as a variable in SEM
}
