% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsem_summary.R
\name{summary.networksem}
\alias{summary.networksem}
\title{Summarize output from networksem}
\usage{
\method{summary}{networksem}(res)
}
\arguments{
\item{res}{a networksem output file}
}
\value{
a summary of the networksem output
}
\description{
Summarize output from networksem
}
