% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkFun.R
\name{sem.net}
\alias{sem.net}
\title{Fit a sem model with network data using node statistics as variables. User-specified network statistics will be calculated and used as variables instead of the networks themselves in the SEM.}
\usage{
sem.net(
  model = NULL,
  data = NULL,
  netstats = NULL,
  ordered = NULL,
  sampling.weights = NULL,
  netstats.rescale = TRUE,
  group = NULL,
  cluster = NULL,
  constraints = "",
  WLS.V = NULL,
  NACOV = NULL,
  netstats.options = NULL,
  ...
)
}
\arguments{
\item{model}{a model specified in lavaan model syntax.}

\item{data}{a list containing the observed non-network nodal variables and the network data}

\item{netstats}{a user-specified list of network statistics to be calculated and used in the SEM, e.g., c("degree", "betweenness"), available options include "degree", "betweenness", "closeness", "evcent", "stresscent", and "infocent" from the "sna" package and "ivi", "hubness.score", "spreading.score" and "clusterRank" from the "influential" package}

\item{ordered}{parameter same as "ordered" in the lavaan sem() function; whether to treat data as ordinal}

\item{sampling.weights}{parameter same as "sampling.weights" in the lavaan sem() function; whether to apply weights to data}

\item{netstats.rescale}{a list of logical value indicating whether to rescale network statistics to have mean 0 and sd 1.}

\item{group}{parameter same as "group" in the lavaan sem() function; whether to fit a multigroup model}

\item{cluster}{parameter same as "cluster" in the lavaan sem() function; whether to fit a cluster model}

\item{constraints}{parameter same as "constraints" in the lavaan sem() function; whether to apply constraints to the model}

\item{WLS.V}{parameter same as "WLS.V" in the lavaan sem() function; whether to use WLS.V estimator}

\item{NACOV}{parameter same as "NACOV" in the lavaan sem() function; whether to use NACOV estimator}

\item{netstats.options}{a user-specified named list with element names corresponding to the network statistics names and element values corresponding to other lists. The list corresponding to each network statistics name has element names being the argument names for calculating the network statistics, and values being the argument values, as used in the corresponding functions in the "sna" or "influential" packages. e.g., netstats.options=list("degree"=list("cmode"="freeman"), "closeness"=list("cmode"="undirected"), "clusterRank"=list("directed"=FALSE))}

\item{...}{optional arguments for the sem() function}
}
\value{
the updated model specification with the network statistics as variables and a lavaan object which is the SEM results
}
\description{
Fit a sem model with network data using node statistics as variables. User-specified network statistics will be calculated and used as variables instead of the networks themselves in the SEM.
}
