% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsem_summary.R
\name{summary.networksem}
\alias{summary.networksem}
\title{Summarize output from networksem}
\usage{
summary.networksem(object, ...)
}
\arguments{
\item{object}{a networksem output}

\item{...}{optional arguments}
}
\value{
a summary of the output
}
\description{
Summarize output from networksem
}
