% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netsem_summary.R
\name{summary.networksem}
\alias{summary.networksem}
\title{Summarize output from networksem functions includeing sem.net, sem.net.lsm, sem.net.edge, sem.net.edge.lsm.}
\usage{
\method{summary}{networksem}(object, ...)
}
\arguments{
\item{object}{A networksem output object.}

\item{...}{Optional arguments.}
}
\value{
a summary sheet of the networksem output.
}
\description{
Summarize output from networksem functions includeing sem.net, sem.net.lsm, sem.net.edge, sem.net.edge.lsm.
}
